refactor components and project structure
