+++ src/App.js
@@ -1,8 +1,8 @@
 import React, { useLayoutEffect } from 'react';
 import './App.css';
-import Content from './components/Content/Content';
+import SearchResultsPage from './components/Content/SearchResultsPage';
 import { Header, Footer } from './components/Layout';
-import SongPage from './components/SongPage';
+import SongPage from './components/Content/SongPage';
 import { Grid } from '@material-ui/core';
 import { Route, Switch, useLocation } from 'react-router-dom';
 
@@ -20,7 +20,7 @@ const App = () => {
         <Grid className='side' xs={false} sm={2} item></Grid>
         <Grid className='main' xs={12} sm={8} item container>
           <Switch>
-            <Route exact path='/' component={Content} />
+            <Route exact path='/' component={SearchResultsPage} />
             <Route exact path='/song/:id?' component={SongPage} />
           </Switch>
         </Grid>
