+++ src/components/Content/SearchResultsPage.js
@@ -45,7 +45,14 @@ const Content = (props) => {
         {tracks.length === 0
           ? 'No results'
           : tracks.map((track) => (
-              <Grid key={track.id} item xs={12}>
+              <Grid
+                key={track.id}
+                item
+                xs={12}
+                style={{
+                  minWidth: '60vw',
+                }}
+              >
                 <NavLink to={`/song/${track.id}`}>
                   <TrackCard
                     key={track.id}
